website_api/requirements.txt
asgiref==3.7.2
atomicwrites==1.4.1
attrs==23.2.0
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==41.0.7
Django==5.0.1
django-cors-headers==4.3.1
django-phonenumber-field==7.3.0
django-sslserver==0.22
djangorestframework==3.14.0
fuzzywuzzy==0.18.0
greenlet==3.0.3
gunicorn==21.2.0
haversine==2.8.1
idna==3.6
packaging==23.2
pathlib-mate==1.3.1
phonenumbers==8.13.27
prettytable==3.9.0
psycopg2==2.9.9
pycparser==2.21
pyOpenSSL==23.3.0
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0
sib-api-v3-sdk==7.6.0
six==1.16.0
SQLAlchemy==1.4.51
sqlalchemy-mate==1.4.28.4
sqlparse==0.4.4
tzdata==2023.4
urllib3==2.1.0
uszipcode==1.0.1
wcwidth==0.2.13
whitenoise==6.6.0
